+++ packages/core/src/reducers/CalendarDataManager.ts
@@ -337,6 +337,7 @@ export class CalendarDataManager {
         // hack
         state.eventSources = data.eventSources = reduceEventSourcesNewTimeZone(data.eventSources, state.dateProfile, data)
         state.eventStore = data.eventStore = rezoneEventStoreDates(data.eventStore, oldData.dateEnv, data.dateEnv)
+        state.renderableEventStore = data.renderableEventStore = rezoneEventStoreDates(data.renderableEventStore, oldData.dateEnv, data.dateEnv)
       }
 
       for (let optionName in changeHandlers) {